+++ src/components/cells/common.tsx
@@ -71,7 +71,7 @@ export function renderChildrenWithDivider(
       {index < React.Children.count(children) - 1 && (
         <View
           style={{
-            width: '95%',
+            width: '94%',
             alignSelf: 'center',
             borderBottomWidth: cellElementOptions.borderWidth,
             borderColor: cellElementOptions.borderColor,